Move buildSplineTemplate() to frameProcessor.py
